Control de entradas y salidas de productos en albaranes

+++ app/controllers/albaran_lineas_controller.rb
@@ -40,18 +40,22 @@ class AlbaranLineasController < ApplicationController
     end    
   end
 
-  def editar_linea
-    @albaranlinea = params[:id] ?  AlbaranLinea.find(params[:id]) : AlbaranLinea.new
-    @albaran = Albaran.find(params[:albaran_id])
-    render :update do |page|
-      page.replace_html params[:update] + "_linea_editar", :partial => "editar_linea"
-    end
-  end
+  #def editar_linea
+  #  @albaranlinea = params[:id] ?  AlbaranLinea.find(params[:id]) : AlbaranLinea.new
+  #  @albaran = Albaran.find(params[:albaran_id])
+  #  render :update do |page|
+  #    page.replace_html params[:update] + "_linea_editar", :partial => "editar_linea"
+  #  end
+  #end
 
   def asignar_linea
-    params[:albaranlinea][:producto_id] = params[:producto][:id]
-    albaranlinea = params[:id] ? AlbaranLinea.find(params[:id]) : AlbaranLinea.new
-    albaranlinea.update_attributes params[:albaranlinea]
+    albaran = Albaran.find_by_id params[:albaranlinea][:albaran_id]
+    if albaran && !albaran.cerrado
+      params[:albaranlinea][:producto_id] = params[:producto][:id]
+      albaran = Albaran.find_by_id params[:albaranlinea][:albaran_id]
+      albaranlinea = params[:id] ? AlbaranLinea.find(params[:id]) : AlbaranLinea.new
+      albaranlinea.update_attributes params[:albaranlinea]
+    end
     #flash[:error] = albaranlinea
     #@albaran_lineas = Albaran.find(albaranlinea.albaran_id).albaran_lineas
     #redirect_to :action => :lineas, :albaran_id => albaranlinea.albaran_id, :update => params[:update]
@@ -62,8 +66,9 @@ class AlbaranLineasController < ApplicationController
   end
 
   def eliminar_linea
+    albaran = Albaran.find_by_id params[:albaran_id]
     albaranlinea = AlbaranLinea.find_by_id params[:id] 
-    if albaranlinea 
+    if albaranlinea && albaran && !albaran.cerrado
       albaranlinea.destroy
       #flash[:error] = albaranlinea
       #@albaran_lineas = Albaran.find(params[:albaran_id]).albaran_lineas
